+++ CHANGELOG.md
@@ -68,3 +68,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Human Readable to bridgehead
 - Show minimal information in project form
 - Show User TODOs
+- Extended explanations